requirements.txt
requests
PyYAML==5.1
